[FEAT] 훅 분리(#110)
